+++ server/paymentProviders/opencollective/collective.ts
@@ -1,8 +1,8 @@
 import { TransactionKind } from '../../constants/transaction-kind';
 import { TransactionTypes } from '../../constants/transactions';
-import { getRefundableAmountFromCollectiveInHostCurrency } from '../../graphql/common/transactions';
 import { getFxRate, roundCentsAmount } from '../../lib/currency';
 import { createRefundTransaction, getHostFee, getHostFeeSharePercent, getPlatformTip } from '../../lib/payments';
+import { getRefundableAmountFromCollectiveInHostCurrency } from '../../lib/refunds';
 import { formatCurrency } from '../../lib/utils';
 import models, { type PaymentMethod } from '../../models';
 import type { BasePaymentProviderService } from '../types';
